refactor: `CoinlistView` Scope, 파일 분리
