+++ CryptocurrencyExchange/CryptocurrencyExchange/Sources/Presentations/Scene/CoinList/CoinListEnviroment.swift
@@ -1,0 +1,13 @@
+//
+//  CoinListEnviroment.swift
+//  CryptocurrencyExchange
+//
+//  Created by 김정상 on 2022/03/12.
+//
+
+import Foundation
+
+struct CoinListEnvironment {
+    let coinListUseCase: () -> CoinListUseCaseProtocol
+    let toastClient: ToastClient
+}
